Move debug and error into a messages module.
