+++ lib/events.js
@@ -3,6 +3,8 @@
 // Event Catalogue.
 //
 // To add an event to the catalogue, look below the definition of `events`.
+import { error } from './messages';
+
 const events = {
   // All activated events. (To be constructed below.)
   dict: {},
@@ -17,7 +19,7 @@ events.add = function add(event) {
 // Gets an event by name. Throws an error if not found.
 events.get = function get(eventName) {
   const event = events.dict[eventName];
-  if (!event) { throw new Error(`Event not recognised: ${eventName}`); }
+  if (!event) { error(`Event not recognised: ${eventName}`); }
   return event;
 };
