+++ lib/events/event.js
@@ -12,21 +12,23 @@
 //
 // trigger = function(params) { ... }
 //   A function called when the event conditions are met.
+import { error } from '../messages';
+
 export default function event(opts) {
   if (!opts.name) {
-    throw new Error(`Event defined without a name!\n${opts}`);
+    error(`Event defined without a name!\n${opts}`);
   }
 
   if (!opts.attach) {
-    throw new Error(`Event ${opts.name} defined without an attach() method!`);
+    error(`Event ${opts.name} defined without an attach() method!`);
   }
 
   if (!opts.trigger) {
-    throw new Error(`Event ${opts.name} defined without a trigger() method!`);
+    error(`Event ${opts.name} defined without a trigger() method!`);
   }
 
   if (!opts.detach) {
-    throw new Error(`Event ${opts.name} defined without a detach() method!`);
+    error(`Event ${opts.name} defined without a detach() method!`);
   }
 
   return {
