+++ lib/messages.js
@@ -5,8 +5,16 @@
 
 import config from './config';
 
-export default function debug(message) {
+// To be used for debug messages.
+// Checks the config to see if debug messages are enabled.
+export function debug(message) {
   if (config.DEBUG) {
     console.log(`[PING] ${message}`); // eslint-disable-line no-console
   }
+}
+
+// Throw an error.
+// TODO: Maybe we want to do something else?
+export function error(message) {
+  throw new Error(`[PING] ${message}`);
 }
