+++ lib/reevoo-ping.js
@@ -5,6 +5,7 @@
 require('./polyfill/array');
 require('./snowplow');
 import eventCatalogue from './events';
+import { error } from './messages';
 
 const reevooPing = {
   // By default, track nothing.
