+++ lib/snowplow.js
@@ -5,7 +5,7 @@
 import internalSnowplow from './snowplow/bootstrap';
 import snowplowConfig from './snowplow/config';
 import config from './config';
-import debug from './debug';
+import { debug } from './messages';
 
 // Initialise the tracker
 internalSnowplow('newTracker', 'cf', config.COLLECTOR_URI, snowplowConfig);
