+++ lib/snowplow/bootstrap.js
@@ -3,7 +3,7 @@
 // Load Snowplow into the window.
 // Adapted from https://github.com/snowplow/snowplow-javascript-tracker/blob/master/tags/tag.js
 
-import debug from '../debug';
+import { debug } from '../messages';
 
 // We could use `window[snowplowFunctionName]` everywhere, but that's not very ES6.
 // Take the export instead.
