--- spec/debug.spec.js
@@ -1,7 +1,0 @@
-import debug from 'lib/debug';
-
-describe('lib/debug', () => {
-  it('is simple enough not to have tests', () => {
-    expect(debug).toBeDefined();
-  });
-});
