+++ spec/messages.spec.js
@@ -1,0 +1,7 @@
+import * as messages from 'lib/messages';
+
+describe('lib/messages', () => {
+  it('is simple enough not to have tests', () => {
+    expect(messages).toBeDefined();
+  });
+});
